Introduce add_custom_yaml_representer as API
