+++ tests/test_data_simple/test_custom_object.yml
@@ -1,0 +1,3 @@
+scalar:
+  unit: m
+  value: 10
